chore: move nock utils and seperate tests
